Revert "Rename dom-for in file paths to domFor"
This reverts commit 5ab2752603526b2eb7368d0f28370a53a8183477.
